Added user roles and control accordint to them

+++ resources/views/home.blade.php
@@ -11,7 +11,9 @@
                         @if(auth()->user()->pasirinkta_tema) {{-- Jei studentas turi tema, tik tada gali jos atsisakyti --}}
                             <a href="/tema/abandon" class="btn btn-primary">Atsisakyti temos</a>
                         @endif
-                        <a href="/tema/insert" class="btn btn-primary">Pridėti</a>
+                        @if(auth()->user()->isAdmin())
+                            <a href="/tema/insert" class="btn btn-primary">Pridėti</a>
+                        @endif
                     </div>
                 </span>
 
@@ -35,8 +37,12 @@
                             <td>{{ $tema->stud_limitas - $tema->pasirinkusieji }}</td>
                             <td>
                                 <a href="/tema/{{$tema->id}}" class="btn btn-primary">Detaliau</a>
-                                <a href="/tema/edit/{{$tema->id}}" class="btn btn-primary">Redaguoti</a>
-                                <a href="/tema/delete/{{$tema->id}}" class="btn btn-danger">Šalinti</a>
+                                @if(auth()->user()->isAdmin())
+                                    <a href="/tema/edit/{{$tema->id}}" class="btn btn-primary">Redaguoti</a>
+                                @endif
+                                @if(auth()->user()->isAdmin())
+                                    <a href="/tema/delete/{{$tema->id}}" class="btn btn-danger">Šalinti</a>
+                                @endif
                                 @if($tema->stud_limitas - $tema->pasirinkusieji > 0 && !auth()->user()->pasirinkta_tema)
                                     <a href="/tema/choose/{{$tema->id}}" class="btn btn-success">Rinktis</a>  {{-- Galima rinktis tik atitikus salygas --}}
                                 @endif
